Functions to generate and classify blocks of text
This uses the regular expression which splits a piece of text into a
list of alternating alphanumeric and non-alphanumeric strings (each such
string is referred to as a block). Then derive a coding for each block
based on whether it is pure alpha, pure numeric, whitespace, contains a
dash, or has other non-alphanumeric characters. These functions will be
used to understand the variations in the structure of the product model
field.