Remove suppress_prod_code_if_equal_to_model
Remove BaseMasterTemplateBuilder.suppress_prod_code_if_equal_to_model as
it didn't work as intended. This is because there are various filters on
previous Listing Matcher Templates, so sometimes a prod code matches
model, but the model templates have been ignored. So the prod code
template can still be useful.