Treat x4 pattern as 'xn' not 'an' to fix Kiss X4
The Kiss X4 was matching a product with a x4 optical zoom. To fix this
issue, treat a block with a single 'x' as having a classification of
'x'. Convert it back to a 'c' (consonant) if it is preceded by a dash
(hence a product code), or if it is not succeeded by a number.